+++ src/main/java/shopping/domain/member/MemberPassword.java
@@ -1,4 +1,4 @@
-package shopping.domain;
+package shopping.domain.member;
 
 import java.util.Objects;
 import javax.persistence.Embeddable;
