+++ src/main/java/shopping/repository/MemberRepository.java
@@ -3,8 +3,8 @@ package shopping.repository;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import shopping.domain.Member;
-import shopping.domain.MemberEmail;
+import shopping.domain.member.Member;
+import shopping.domain.member.MemberEmail;
 
 @Repository
 public interface MemberRepository extends JpaRepository<Member, Long> {
