+++ src/test/java/shopping/domain/CartProductCountTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import shopping.domain.cart.CartProductCount;
 import shopping.exception.ShoppingException;
 
 class CartProductCountTest {
