+++ src/test/java/shopping/domain/MemberEmailTest.java
@@ -8,6 +8,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
+import shopping.domain.member.MemberEmail;
 import shopping.exception.ShoppingException;
 
 class MemberEmailTest {
