+++ src/test/java/shopping/domain/MemberTest.java
@@ -4,6 +4,9 @@ import static org.assertj.core.api.Assertions.assertThatCode;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import shopping.domain.member.Member;
+import shopping.domain.member.MemberEmail;
+import shopping.domain.member.MemberPassword;
 
 class MemberTest {
