+++ src/test/java/shopping/domain/ProductImageTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
+import shopping.domain.product.ProductImage;
 import shopping.exception.ShoppingException;
 
 class ProductImageTest {
