+++ src/test/java/shopping/repository/MemberRepositoryTest.java
@@ -7,8 +7,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import shopping.domain.Member;
-import shopping.domain.MemberEmail;
+import shopping.domain.member.Member;
+import shopping.domain.member.MemberEmail;
 
 @DataJpaTest
 class MemberRepositoryTest {
